+++ packages/extension-base/src/services/swap-service/index.ts
@@ -106,6 +106,8 @@ export class SwapService implements ServiceWithProcessInterface, StoppableServic
       selectedQuote: swapQuoteResponse.optimalQuote
     });
 
+    console.log('optimalProcess', optimalProcess);
+
     return {
       process: optimalProcess,
       quote: swapQuoteResponse