+++ packages/extension-koni-ui/src/Popup/Confirmations/index.tsx
@@ -103,6 +103,16 @@ const Component = function ({ className }: Props) {
       }
     }
 
+    if (confirmation.item.isInternal && confirmation.type !== 'connectWCRequest') {
+      return (
+        <TransactionConfirmation
+          closeAlert={closeAlert}
+          confirmation={confirmation}
+          openAlert={openAlert}
+        />
+      );
+    }
+
     switch (confirmation.type) {
       case 'addNetworkRequest':
         return <AddNetworkConfirmation request={confirmation.item as ConfirmationDefinitions['addNetworkRequest'][0]} />;
@@ -140,16 +150,6 @@ const Component = function ({ className }: Props) {
         );
       case 'notSupportWCRequest':
         return (<NotSupportWCConfirmation request={confirmation.item as WalletConnectNotSupportRequest} />);
-    }
-
-    if (confirmation.item.isInternal) {
-      return (
-        <TransactionConfirmation
-          closeAlert={closeAlert}
-          confirmation={confirmation}
-          openAlert={openAlert}
-        />
-      );
     }
 
     return null;
@@ -230,7 +230,7 @@ const Component = function ({ className }: Props) {
         case ExtrinsicType.TOKEN_APPROVE:
           return t('Token approve');
         case ExtrinsicType.SWAP:
-          return t('Swap confirm');
+          return t('Swap confirmation');
         case ExtrinsicType.CROWDLOAN:
         case ExtrinsicType.EVM_EXECUTE:
         case ExtrinsicType.UNKNOWN: