+++ packages/extension-koni-ui/src/Popup/Home/History/Detail/parts/Layout.tsx
@@ -1,9 +1,11 @@
 // Copyright 2019-2022 @subwallet/extension-koni-ui authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+import { ExtrinsicType } from '@subwallet/extension-base/background/KoniTypes';
 import { MetaInfo } from '@subwallet/extension-koni-ui/components';
 import { HistoryStatusMap, TxTypeNameMap } from '@subwallet/extension-koni-ui/constants';
 import { useSelector } from '@subwallet/extension-koni-ui/hooks';
+import SwapLayout from '@subwallet/extension-koni-ui/Popup/Home/History/Detail/parts/SwapLayout';
 import { ThemeProps, TransactionHistoryDisplayItem } from '@subwallet/extension-koni-ui/types';
 import { formatHistoryDate, isAbleToShowFee, toShort } from '@subwallet/extension-koni-ui/utils';
 import CN from 'classnames';
@@ -25,6 +27,12 @@ const Component: React.FC<Props> = (props: Props) => {
   const { t } = useTranslation();
 
   const { language } = useSelector((state) => state.settings);
+
+  if (data.type === ExtrinsicType.SWAP) {
+    return (
+      <SwapLayout data={data} />
+    );
+  }
 
   return (
     <MetaInfo className={CN(className)}>