+++ packages/extension-koni-ui/src/constants/transaction.ts
@@ -3,7 +3,7 @@
 
 import { ExtrinsicType, StakingType } from '@subwallet/extension-base/background/KoniTypes';
 import { detectTranslate } from '@subwallet/extension-base/utils';
-import { CancelUnStakeParams, ClaimRewardParams, EarnParams, SendNftParams, StakeParams, TransactionFormBaseProps, TransferParams, UnStakeParams, WithdrawParams } from '@subwallet/extension-koni-ui/types';
+import { CancelUnStakeParams, ClaimRewardParams, EarnParams, SendNftParams, StakeParams, SwapParams, TransactionFormBaseProps, TransferParams, UnStakeParams, WithdrawParams } from '@subwallet/extension-koni-ui/types';
 
 import { ALL_KEY } from './common';
 
@@ -130,4 +130,12 @@ export const DEFAULT_CLAIM_REWARD_PARAMS: ClaimRewardParams = {
   ...DEFAULT_TRANSACTION_PARAMS,
   slug: '',
   bondReward: false
+};
+
+export const DEFAULT_SWAP_PARAMS: SwapParams = {
+  ...DEFAULT_TRANSACTION_PARAMS,
+  fromAmount: '',
+  fromTokenSlug: '',
+  toTokenSlug: '',
+  defaultSlug: ''
 };