+++ packages/extension-koni-ui/src/reducer/index.ts
@@ -3,3 +3,4 @@
 
 export * from './signing';
 export * from './earning';
+export * from './swap';